move logic related to opening and initializing handle from FileStreamHelpers to SafeFileHandle itself
